Breaking out ExtensionModel.Xml into smaller files. #189
